Lpnamer: create MasterGeneration class
